requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-core==1.35.1
azure-identity==1.25.0
azure-storage-blob==12.26.0
banks==2.2.0
beautifulsoup4==4.13.4
boto3==1.38.46
botocore==1.38.46
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.115.14
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
limits==5.4.0
llama-cloud==0.1.41
llama-cloud-services==0.6.65
llama-index-core==0.14.1
llama-index-embeddings-azure-openai==0.4.1
llama-index-embeddings-openai==0.5.1
llama-index-instrumentation==0.4.0
llama-index-llms-azure-openai==0.4.1
llama-index-llms-openai==0.5.6
llama-index-workflows==2.1.0
llama-parse==0.6.65
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.1
openai==1.93.0
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.2
PyMuPDF==1.26.1
pypdf==6.1.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.0
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.46.2
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.35.0
validate-email-address==1
wrapt==1.17.2
yarl==1.20.1
gunicorn==23.0.0
